moved specific ncanda and hivalc functions to respective mapping files
